appsettings

+++ BerkutPoiService/Models/PointOfInterest.cs
@@ -1,20 +1,13 @@
 ﻿using Azure;
 using Azure.Data.Tables;
-using BerkutPoiService.Options;
-using Microsoft.Extensions.Options;
 using System;
-using System.Collections.Concurrent;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace BerkutPoiService.Models
 {
     public class PointOfInterest : ITableEntity
     {
-        public decimal Lat { get; set; }
-        public decimal Long { get; set; }
+        public double Lat { get; set; }
+        public double Long { get; set; }
         public string Name { get; set; }
         public string Content { get; set; }
         public string GeoHash { get; set; }
